Fix build errors using clang on windows with JIT
