+++ lib/Backend/GlobHashTable.h
@@ -276,7 +276,7 @@ public:
     {
         for (uint i = 0; i < this->tableSize; i++)
         {
-            SListBase<HashBucket>::Iterator iter2(&this2->table[i]);
+            _TYPENAME SListBase<HashBucket>::Iterator iter2(&this2->table[i]);
             iter2.Next();
             FOREACH_SLISTBASE_ENTRY_EDITING(HashBucket, bucket, &this->table[i], iter)
             {
@@ -308,7 +308,7 @@ public:
     {
         for (uint i = 0; i < this->tableSize; i++)
         {
-            SListBase<HashBucket>::Iterator iter2(&this2->table[i]);
+            _TYPENAME SListBase<HashBucket>::Iterator iter2(&this2->table[i]);
             iter2.Next();
             FOREACH_SLISTBASE_ENTRY_EDITING((HashBucket), bucket, &this->table[i], iter)
             {
@@ -347,7 +347,7 @@ public:
 
         for (uint i = 0; i < this->tableSize; i++)
         {
-            this->table[i].CopyTo<HashBucket::Copy>(this->alloc, newTable->table[i]);
+            this->table[i].template CopyTo<HashBucket::Copy>(this->alloc, newTable->table[i]);
         }
 #if PROFILE_DICTIONARY
         if (stats)
