+++ lib/Backend/GlobOptFields.cpp
@@ -2659,7 +2659,7 @@ GlobOpt::OptNewScObject(IR::Instr** instrPtr, Value* srcVal)
 
     if (!instr->IsNewScObjectInstr())
     {
-        return false;
+        return nullptr;
     }
 
     bool isCtorInlined = instr->m_opcode == Js::OpCode::NewScObjectNoCtor;
