+++ lib/Backend/JnHelperMethod.cpp
@@ -155,7 +155,7 @@ GetMethodAddress(IR::HelperCallOpnd* opnd)
 
 // TODO:  Remove this define once makes it into WINNT.h
 #ifndef DECLSPEC_GUARDIGNORE
-#if (_MSC_FULL_VER >= 170065501)
+#if (_MSC_FULL_VER >= 170065501) && !defined(__clang__)
 #define DECLSPEC_GUARDIGNORE  __declspec(guard(ignore))
 #else
 #define DECLSPEC_GUARDIGNORE
