+++ lib/Runtime/Language/DynamicProfileInfo.h
@@ -533,12 +533,14 @@ namespace Js
         static CriticalSection s_csOutput;
         template <typename T>
         static void WriteData(T data, FILE * file);
+#if defined(_MSC_VER) && !defined(__clang__)
         template <>
         static void WriteData<char16 const *>(char16 const * sz, FILE * file);
         template <>
         static void WriteData<FunctionInfo *>(FunctionInfo * functionInfo, FILE * file); // Not defined, to prevent accidentally writing function info
         template <>
         static void WriteData<FunctionBody *>(FunctionBody * functionInfo, FILE * file);
+#endif
         template <typename T>
         static void WriteArray(uint count, T * arr, FILE * file);
 #endif
